requirements.txt
fastapi
uvicorn
motor
pandas
scikit-learn
python-dotenv